Code refactoring and cleanup
