+++ src/split-button/interfaces.ts
@@ -7,21 +7,65 @@ import { BaseComponentProps } from '../internal/base-component';
 import { ButtonProps } from '../button/interfaces';
 import { IconProps } from '../icon/interfaces';
 
-// TODO: api-docs
 export interface SplitButtonProps extends BaseComponentProps {
+  /**
+   * A list of split button items.
+   * The supported item types are: `button` and `button-dropdown`.
+   *
+   * The `button-dropdown` item is only allowed as the very last item in the list.
+   *
+   * The following properties are supported across by utility types:
+   *
+   * * `id` (string) - The type of the item. Can be `button` or `button-dropdown`.
+   * * `type` (string) - ...
+   * * `ariaLabel` (string) - ...
+   * * `disabled` (string) - ...
+   * * `loading` (string) - ...
+   * * `loadingText` (string) - ...
+   *
+   * ### button
+   *
+   * * `text` (string) - ...
+   * * `href` (string) - ...
+   * * `target` (string) - ...
+   * * `rel` (string) - ...
+   * * `download` (string) - ...
+   * * `external` (string) - ...
+   * * `onClick` (() => void) - Specifies the event handler called when the utility is clicked.
+   * * `onFollow` (...) - ...
+   * * `iconName` (string) - ...
+   * * `iconAlt` (string) - ...
+   * * `iconUrl` (string) - ...
+   * * `iconSvg` (string) - ...
+   *
+   * ### button-dropdown
+   *
+   * * `items` (ButtonDropdownProps.Items) - An array of dropdown items. This follows the same structure as the `items` property of the [button dropdown component](/components/button-dropdown).
+   * * `onItemClick` (NonCancelableEventHandler<ButtonDropdownProps.ItemClickDetails>) - Specifies the event handler called when a dropdown item is selected.
+   * * `onItemFollow` (NonCancelableEventHandler<ButtonDropdownProps.ItemClickDetails>) - Specifies the event handler called when a dropdown item is selected.
+   */
   items: ReadonlyArray<SplitButtonProps.Item>;
 
+  /**
+   * ...
+   */
   variant?: SplitButtonProps.Variant;
 
+  /**
+   * ...
+   */
   expandToViewport?: boolean;
 
+  /**
+   * ...
+   */
   expandableGroups?: boolean;
 }
 
 export namespace SplitButtonProps {
   export type Variant = 'normal' | 'primary';
 
-  export type Item = ButtonItem | LinkItem | ButtonDropdownItem;
+  export type Item = ButtonItem | ButtonDropdownItem;
 
   export interface BaseItem {
     id: string;
@@ -31,21 +75,8 @@ export namespace SplitButtonProps {
     loadingText?: string;
   }
 
-  export interface BaseItemIcon {
-    iconName?: IconProps.Name;
-    iconAlt?: string;
-    iconUrl?: string;
-    iconSvg?: string;
-  }
-
-  export interface ButtonItem extends BaseItem, BaseItemIcon {
+  export interface ButtonItem extends BaseItem {
     type: 'button';
-    text?: string;
-    onClick?: CancelableEventHandler;
-  }
-
-  export interface LinkItem extends BaseItem, BaseItemIcon {
-    type: 'link';
     text?: string;
     href?: string;
     target?: string;
@@ -54,6 +85,10 @@ export namespace SplitButtonProps {
     external?: boolean;
     onClick?: CancelableEventHandler;
     onFollow?: CancelableEventHandler<ButtonProps.FollowDetail>;
+    iconName?: IconProps.Name;
+    iconAlt?: string;
+    iconUrl?: string;
+    iconSvg?: string;
   }
 
   export interface ButtonDropdownItem extends BaseItem {
