+++ src/test-utils/dom/split-button/index.ts
@@ -3,7 +3,6 @@
 
 import { ComponentWrapper } from '@cloudscape-design/test-utils-core/dom';
 import splitButtonStyles from '../../../split-button/styles.selectors.js';
-import splitButtonPartsStyles from '../../../split-button/components/styles.selectors.js';
 import ButtonWrapper from '../button';
 import ButtonDropdownWrapper from '../button-dropdown';
 
@@ -11,13 +10,13 @@ export default class SplitButtonDropdownWrapper extends ComponentWrapper {
   static rootSelector: string = splitButtonStyles.root;
 
   findItems(): Array<SplitButtonItemWrapper> {
-    return this.findAll(`.${splitButtonPartsStyles['segment-wrapper']}`).map(
+    return this.findAll(`.${splitButtonStyles['segment-wrapper']}`).map(
       w => new SplitButtonItemWrapper(w.getElement())
     );
   }
 
   findItemById(id: string): null | SplitButtonItemWrapper {
-    const itemSelector = `.${splitButtonPartsStyles['segment-wrapper']}[data-testid="${id}"]`;
+    const itemSelector = `.${splitButtonStyles['segment-wrapper']}[data-testid="${id}"]`;
     const segment = this.find(itemSelector);
     return segment ? new SplitButtonItemWrapper(segment.getElement()) : null;
   }
